backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.2.0
django-filter==24.2
drf-yasg==1.21.7
redis==5.0.1
celery==5.3.6
channels==4.0.0
channels-redis==4.2.0
requests==2.31.0
googlemaps==4.10.0
twilio==9.0.0
sendgrid==6.11.0
django-storages==1.14.2
boto3==1.34.51
django-ckeditor==6.7.0
django-import-export==4.0.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3
djangorestframework-gis==1.0